chore(model): RPC conversion code isolation (S-V)
