+++ crates/api-model/src/site_explorer/mod.rs
@@ -46,20 +46,8 @@ use crate::switch::switch_id;
 #[derive(Clone, Debug, Default)]
 pub struct ExploredEndpointSearchFilter {}
 
-impl From<rpc::site_explorer::ExploredEndpointSearchFilter> for ExploredEndpointSearchFilter {
-    fn from(_filter: rpc::site_explorer::ExploredEndpointSearchFilter) -> Self {
-        ExploredEndpointSearchFilter {}
-    }
-}
-
 #[derive(Clone, Debug, Default)]
 pub struct ExploredManagedHostSearchFilter {}
-
-impl From<rpc::site_explorer::ExploredManagedHostSearchFilter> for ExploredManagedHostSearchFilter {
-    fn from(_filter: rpc::site_explorer::ExploredManagedHostSearchFilter) -> Self {
-        ExploredManagedHostSearchFilter {}
-    }
-}
 
 /// Data that we gathered about a particular endpoint during site exploration
 /// This data is stored as JSON in the Database. Therefore the format can
@@ -487,61 +475,6 @@ pub struct SystemStatus {
     pub state: String,
 }
 
-impl From<SystemStatus> for rpc::site_explorer::SystemStatus {
-    fn from(status: SystemStatus) -> Self {
-        rpc::site_explorer::SystemStatus {
-            health: status.health,
-            health_rollup: status.health_rollup,
-            state: status.state,
-        }
-    }
-}
-
-impl From<PCIeDevice> for rpc::site_explorer::PcIeDevice {
-    fn from(device: PCIeDevice) -> Self {
-        rpc::site_explorer::PcIeDevice {
-            description: device.description,
-            firmware_version: device.firmware_version,
-            gpu_vendor: device.gpu_vendor,
-            id: device.id,
-            manufacturer: device.manufacturer,
-            name: device.name,
-            part_number: device.part_number,
-            serial_number: device.serial_number,
-            status: device.status.map(Into::into),
-        }
-    }
-}
-
-impl From<ExploredEndpoint> for rpc::site_explorer::ExploredEndpoint {
-    fn from(endpoint: ExploredEndpoint) -> Self {
-        rpc::site_explorer::ExploredEndpoint {
-            address: endpoint.address.to_string(),
-            report: Some(endpoint.report.into()),
-            report_version: endpoint.report_version.to_string(),
-            exploration_requested: endpoint.exploration_requested,
-            preingestion_state: format!("{:?}", endpoint.preingestion_state),
-            last_redfish_bmc_reset: endpoint
-                .last_redfish_bmc_reset
-                .map(|time| time.to_string())
-                .unwrap_or_else(|| "no timestamp available".to_string()),
-            last_ipmitool_bmc_reset: endpoint
-                .last_ipmitool_bmc_reset
-                .map(|time| time.to_string())
-                .unwrap_or_else(|| "no timestamp available".to_string()),
-            last_redfish_reboot: endpoint
-                .last_redfish_reboot
-                .map(|time| time.to_string())
-                .unwrap_or_else(|| "no timestamp available".to_string()),
-            last_redfish_powercycle: endpoint
-                .last_redfish_powercycle
-                .map(|time| time.to_string())
-                .unwrap_or_else(|| "no timestamp available".to_string()),
-            pause_remediation: endpoint.pause_remediation,
-        }
-    }
-}
-
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "PascalCase")]
 pub struct ExploredDpu {
@@ -553,15 +486,6 @@ pub struct ExploredDpu {
 
     #[serde(skip)]
     pub report: Arc<EndpointExplorationReport>,
-}
-
-impl From<&ExploredDpu> for rpc::site_explorer::ExploredDpu {
-    fn from(dpu: &ExploredDpu) -> Self {
-        rpc::site_explorer::ExploredDpu {
-            bmc_ip: dpu.bmc_ip.to_string(),
-            host_pf_mac_address: dpu.host_pf_mac_address.map(|m| m.to_string()),
-        }
-    }
 }
 
 impl ExploredDpu {
@@ -736,27 +660,6 @@ mod serialize_option_display {
         match value {
             None => Ok(None),
             Some(value) => Ok(Some(T::from_str(&value).map_err(de::Error::custom)?)),
-        }
-    }
-}
-
-impl From<ExploredManagedHost> for rpc::site_explorer::ExploredManagedHost {
-    fn from(host: ExploredManagedHost) -> Self {
-        rpc::site_explorer::ExploredManagedHost {
-            host_bmc_ip: host.host_bmc_ip.to_string(),
-            dpus: host
-                .dpus
-                .iter()
-                .map(rpc::site_explorer::ExploredDpu::from)
-                .collect(),
-            dpu_bmc_ip: host
-                .dpus
-                .first()
-                .map_or("".to_string(), |d| d.bmc_ip.to_string()),
-            host_pf_mac_address: host
-                .dpus
-                .first()
-                .and_then(|d| d.host_pf_mac_address.map(|m| m.to_string())),
         }
     }
 }
@@ -1079,15 +982,6 @@ impl EndpointExplorationReport {
     }
 }
 
-impl From<SiteExplorationReport> for rpc::site_explorer::SiteExplorationReport {
-    fn from(report: SiteExplorationReport) -> Self {
-        rpc::site_explorer::SiteExplorationReport {
-            endpoints: report.endpoints.into_iter().map(Into::into).collect(),
-            managed_hosts: report.managed_hosts.into_iter().map(Into::into).collect(),
-        }
-    }
-}
-
 /// Describes errors that might have been encountered during exploring an endpoint
 #[derive(thiserror::Error, PartialEq, Eq, Clone, Debug, Serialize, Deserialize)]
 #[serde(tag = "Type", rename_all = "PascalCase")]
@@ -1260,17 +1154,6 @@ pub struct ComputerSystemAttributes {
     pub is_infinite_boot_enabled: Option<bool>,
 }
 
-impl From<ComputerSystemAttributes> for rpc::site_explorer::ComputerSystemAttributes {
-    fn from(attributes: ComputerSystemAttributes) -> Self {
-        rpc::site_explorer::ComputerSystemAttributes {
-            nic_mode: attributes.nic_mode.map(|a| match a {
-                NicMode::Nic => rpc::site_explorer::NicMode::Nic.into(),
-                NicMode::Dpu => rpc::site_explorer::NicMode::Dpu.into(),
-            }),
-        }
-    }
-}
-
 /// `ComputerSystem` definition. Matches redfish definition
 #[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize)]
 #[serde(rename_all = "PascalCase")]
@@ -1318,41 +1201,6 @@ impl ComputerSystem {
     }
 }
 
-impl From<ComputerSystem> for rpc::site_explorer::ComputerSystem {
-    fn from(system: ComputerSystem) -> Self {
-        rpc::site_explorer::ComputerSystem {
-            id: system.id,
-            manufacturer: system.manufacturer,
-            model: system.model,
-            serial_number: system.serial_number,
-            ethernet_interfaces: system
-                .ethernet_interfaces
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-            attributes: Some(rpc::site_explorer::ComputerSystemAttributes::from(
-                system.attributes,
-            )),
-            pcie_devices: system.pcie_devices.into_iter().map(Into::into).collect(),
-            power_state: rpc::site_explorer::PowerState::from(system.power_state) as _,
-            boot_order: system.boot_order.map(|order| order.into()),
-        }
-    }
-}
-
-impl From<PowerState> for rpc::site_explorer::PowerState {
-    fn from(state: PowerState) -> Self {
-        match state {
-            PowerState::Off => rpc::site_explorer::PowerState::Off,
-            PowerState::On => rpc::site_explorer::PowerState::On,
-            PowerState::PoweringOff => rpc::site_explorer::PowerState::PoweringOff,
-            PowerState::PoweringOn => rpc::site_explorer::PowerState::PoweringOn,
-            PowerState::Paused => rpc::site_explorer::PowerState::Paused,
-            PowerState::Unknown => rpc::site_explorer::PowerState::Unknown,
-        }
-    }
-}
-
 #[derive(Debug, Default, Serialize, Deserialize, Clone, Copy, PartialEq, Eq)]
 pub enum PowerState {
     Off,
@@ -1371,19 +1219,6 @@ pub struct Manager {
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub ethernet_interfaces: Vec<EthernetInterface>,
     pub id: String,
-}
-
-impl From<Manager> for rpc::site_explorer::Manager {
-    fn from(manager: Manager) -> Self {
-        rpc::site_explorer::Manager {
-            id: manager.id,
-            ethernet_interfaces: manager
-                .ethernet_interfaces
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-        }
-    }
 }
 
 /// `EthernetInterface` definition. Matches redfish definition
@@ -1458,18 +1293,6 @@ impl FromStr for UefiDevicePath {
     }
 }
 
-impl From<EthernetInterface> for rpc::site_explorer::EthernetInterface {
-    fn from(interface: EthernetInterface) -> Self {
-        rpc::site_explorer::EthernetInterface {
-            id: interface.id,
-            description: interface.description,
-            interface_enabled: interface.interface_enabled,
-            mac_address: interface.mac_address.map(|mac| mac.to_string()),
-            link_status: interface.link_status,
-        }
-    }
-}
-
 /// `Chassis` definition. Matches redfish definition
 #[derive(Clone, PartialEq, Eq, Debug, Default, Serialize, Deserialize)]
 #[serde(rename_all = "PascalCase")]
@@ -1491,23 +1314,6 @@ pub struct Chassis {
     pub revision_id: Option<i32>,
 }
 
-impl From<Chassis> for rpc::site_explorer::Chassis {
-    fn from(chassis: Chassis) -> Self {
-        rpc::site_explorer::Chassis {
-            id: chassis.id,
-            manufacturer: chassis.manufacturer,
-            model: chassis.model,
-            part_number: chassis.part_number,
-            serial_number: chassis.serial_number,
-            network_adapters: chassis
-                .network_adapters
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-        }
-    }
-}
-
 /// `NetworkAdapter` definition. Matches redfish definition
 #[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "PascalCase")]
@@ -1521,31 +1327,11 @@ pub struct NetworkAdapter {
     pub serial_number: Option<String>,
 }
 
-impl From<NetworkAdapter> for rpc::site_explorer::NetworkAdapter {
-    fn from(adapter: NetworkAdapter) -> Self {
-        rpc::site_explorer::NetworkAdapter {
-            id: adapter.id,
-            manufacturer: adapter.manufacturer,
-            model: adapter.model,
-            part_number: adapter.part_number,
-            serial_number: adapter.serial_number,
-        }
-    }
-}
-
 /// `SecureBootStatus` definition.
 #[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "PascalCase")]
 pub struct SecureBootStatus {
     pub is_enabled: bool,
-}
-
-impl From<SecureBootStatus> for rpc::site_explorer::SecureBootStatus {
-    fn from(secure_boot_status: SecureBootStatus) -> Self {
-        rpc::site_explorer::SecureBootStatus {
-            is_enabled: secure_boot_status.is_enabled,
-        }
-    }
 }
 
 /// `LockdownStatus` definition. Matches redfish definition
@@ -1556,33 +1342,12 @@ pub struct LockdownStatus {
     pub message: String,
 }
 
-impl From<LockdownStatus> for rpc::site_explorer::LockdownStatus {
-    fn from(lockdown_status: LockdownStatus) -> Self {
-        rpc::site_explorer::LockdownStatus {
-            status: rpc::site_explorer::InternalLockdownStatus::from(lockdown_status.status) as _,
-            message: lockdown_status.message,
-        }
-    }
-}
-
 #[derive(Debug, Default, Serialize, Deserialize, Clone, Copy, PartialEq, Eq)]
 pub enum InternalLockdownStatus {
     Enabled,
     Partial,
     #[default]
     Disabled,
-}
-
-impl From<InternalLockdownStatus> for rpc::site_explorer::InternalLockdownStatus {
-    fn from(state: InternalLockdownStatus) -> Self {
-        match state {
-            InternalLockdownStatus::Enabled => rpc::site_explorer::InternalLockdownStatus::Enabled,
-            InternalLockdownStatus::Partial => rpc::site_explorer::InternalLockdownStatus::Partial,
-            InternalLockdownStatus::Disabled => {
-                rpc::site_explorer::InternalLockdownStatus::Disabled
-            }
-        }
-    }
 }
 
 /// `Service` definition. Matches redfish definition
@@ -1592,15 +1357,6 @@ pub struct Service {
     pub id: String,
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub inventories: Vec<Inventory>,
-}
-
-impl From<Service> for rpc::site_explorer::Service {
-    fn from(service: Service) -> Self {
-        rpc::site_explorer::Service {
-            id: service.id,
-            inventories: service.inventories.into_iter().map(Into::into).collect(),
-        }
-    }
 }
 
 /// `Inventory` definition. Matches redfish definition
@@ -1613,17 +1369,6 @@ pub struct Inventory {
     pub release_date: Option<String>,
 }
 
-impl From<Inventory> for rpc::site_explorer::Inventory {
-    fn from(inventory: Inventory) -> Self {
-        rpc::site_explorer::Inventory {
-            id: inventory.id,
-            description: inventory.description,
-            version: inventory.version,
-            release_date: inventory.release_date,
-        }
-    }
-}
-
 /// `MachineSetupStatus` definition. Matches redfish definition
 #[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "PascalCase")]
@@ -1632,32 +1377,11 @@ pub struct MachineSetupStatus {
     pub diffs: Vec<MachineSetupDiff>,
 }
 
-impl From<MachineSetupStatus> for rpc::site_explorer::MachineSetupStatus {
-    fn from(machine_setup_status: MachineSetupStatus) -> Self {
-        rpc::site_explorer::MachineSetupStatus {
-            is_done: machine_setup_status.is_done,
-            diffs: machine_setup_status
-                .diffs
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-        }
-    }
-}
-
 /// `BootOrder` definition.
 #[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "PascalCase")]
 pub struct BootOrder {
     pub boot_order: Vec<BootOption>,
-}
-
-impl From<BootOrder> for rpc::site_explorer::BootOrder {
-    fn from(order: BootOrder) -> Self {
-        rpc::site_explorer::BootOrder {
-            boot_order: order.boot_order.into_iter().map(Into::into).collect(),
-        }
-    }
 }
 
 /// `MachineSetupDiff` definition. Matches redfish definition
@@ -1669,16 +1393,6 @@ pub struct MachineSetupDiff {
     pub actual: String,
 }
 
-impl From<MachineSetupDiff> for rpc::site_explorer::MachineSetupDiff {
-    fn from(machine_setup_diff: MachineSetupDiff) -> Self {
-        rpc::site_explorer::MachineSetupDiff {
-            key: machine_setup_diff.key,
-            expected: machine_setup_diff.expected,
-            actual: machine_setup_diff.actual,
-        }
-    }
-}
-
 /// `BootOption` definition.
 #[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "PascalCase")]
@@ -1687,17 +1401,6 @@ pub struct BootOption {
     pub id: String,
     pub boot_option_enabled: Option<bool>,
     pub uefi_device_path: Option<String>,
-}
-
-impl From<BootOption> for rpc::site_explorer::BootOption {
-    fn from(boot_option: BootOption) -> Self {
-        rpc::site_explorer::BootOption {
-            display_name: boot_option.display_name,
-            id: boot_option.id,
-            boot_option_enabled: boot_option.boot_option_enabled,
-            uefi_device_path: boot_option.uefi_device_path,
-        }
-    }
 }
 
 /// Whether a found/explored machine is in the set of expected machines,
