+++ crates/api-model/src/state_history.rs
@@ -55,42 +55,3 @@ impl<'r> FromRow<'r, PgRow> for StateHistoryRecord {
         })
     }
 }
-
-impl From<StateHistoryRecord> for ::rpc::forge::StateHistoryRecord {
-    fn from(value: StateHistoryRecord) -> ::rpc::forge::StateHistoryRecord {
-        let time = value
-            .time
-            .unwrap_or_else(|| value.state_version.timestamp());
-        ::rpc::forge::StateHistoryRecord {
-            state: value.state,
-            version: value.state_version.version_string(),
-            time: Some(time.into()),
-        }
-    }
-}
-
-impl From<StateHistoryRecord> for ::rpc::forge::MachineEvent {
-    fn from(value: StateHistoryRecord) -> ::rpc::forge::MachineEvent {
-        let time = value
-            .time
-            .unwrap_or_else(|| value.state_version.timestamp());
-        ::rpc::forge::MachineEvent {
-            event: value.state,
-            version: value.state_version.version_string(),
-            time: Some(time.into()),
-        }
-    }
-}
-
-impl From<StateHistoryRecord> for ::rpc::forge::NetworkSegmentStateHistory {
-    fn from(value: StateHistoryRecord) -> ::rpc::forge::NetworkSegmentStateHistory {
-        let time = value
-            .time
-            .unwrap_or_else(|| value.state_version.timestamp());
-        ::rpc::forge::NetworkSegmentStateHistory {
-            state: value.state,
-            version: value.state_version.version_string(),
-            time: Some(time.into()),
-        }
-    }
-}
