+++ crates/api-model/src/trim_table.rs
@@ -19,11 +19,3 @@
 pub enum TrimTableTarget {
     MeasuredBoot,
 }
-
-impl From<rpc::forge::TrimTableTarget> for TrimTableTarget {
-    fn from(target: rpc::forge::TrimTableTarget) -> Self {
-        match target {
-            rpc::forge::TrimTableTarget::MeasuredBoot => TrimTableTarget::MeasuredBoot,
-        }
-    }
-}
